requirements.txt
python-dotenv
pydantic
fastapi
uvicorn
mysql-connector-python
pymongo
redis
strawberry-graphql